requirements.txt
langchain==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langsmith==0.1.144
openai==1.55.0
python-dotenv==1.0.1